backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "gunicorn>=23.0.0",
    "kafka>=1.3.5",
    "neo4j>=5.27.0",
    "numpy>=2.2.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.5",
    "pyspark>=3.5.4",
    "python-decouple>=3.8",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.37",
    "sqlmodel>=0.0.22",
    "streamlit>=1.41.1",
    "uvicorn>=0.34.0",
]
